Move conversion to/from `thir::Pat` out of deconstruct_pat
